Likelihood function for Chow-Lin or Litterman temporal disaggregation.

This function computes the likelihood function used in temporal disaggregation to find the optimal \(\rho\) parameter. It is used in conjunction with disaggregate to estimate the autocorrelation coefficient \(\rho\).

Usage

chowlin_likelihood(Y, X, vcov)

Arguments

Y

The low-frequency response series (an \(n_l \times 1\) matrix).

X

The aggregated high-frequency indicator series (an \(n_l \times p\) matrix).

vcov

Aggregated variance-covariance matrix for the Chow-Lin or Litterman residuals.

Value

    The log-likelihood value for the given parameters.